refactor: split memory module — show/archive/supersede into own files, date helpers to utils/dates (PR#134 review)
